requirements.txt
Flask
pyepics